Refactored some code and moved some classes to the new package, 'utils'
